Added wall breach
Wall breach is a feature that allows for wall reinforcement using walls, but at the cost of big and behemoth biters dealing more damage to the wall structures if they are not reinforced enough

If the concept is appealing it could be expanded, with tiles giving further reinforcements.

Big biters deal extra damage to 1 row walls, and near normal damage to 2 row walls

Behemoth biters deal extra damage to 1 row walls and 2 row walls and regular damage to 3 row walls.

Creates alerts if the walls are being breached.