Update hail_hydra.lua
Fixes an infrequent bug in hail_hydra that has existed a long time.    Currently Factorio does not allow enemy entitities of type 'turret' to target entities of type 'car' (includes cars and tanks).  Doing so logs an error.  If this changes in the future then the condition cause.type ~= "car" below can be removed.