Fixed /strike command range issue
Previously the airstrike was not reaching the corners of the map. There was a default max_range value of 1000 that limited how far the poison capsules could travel. Have set it to 1400 so that it can reach the corners of the map.

THIS NEEDS PLAY TESTING. Please don't merge until I've confirmed I've play tested it.