Updated strike cost
The strike count and cost described by the market was correct but the actual number of capsules sent and the number of capsules removed from the chest was wrong. Bug due to starting index, have now updated the maths to be correct.

Tested and the market description, number of shots fired and the number of capsules removed from the chest are now correct.