Updated outpost pollution to spawn at map centre not at outposts
No one likes or uses the pollution multiplier feature I added.

Feedback suggests that it spawning at outposts makes the cause too obvious, so people complain about it. And it requires that outposts have to be defended.

Yet, people still complain crash site is too easy. 

By moving the pollution spawning to the centre of the base the effect of more outposts is spread over the whole base and will trigger more attacks, but will challenge the whole base defences rather than just a single outpost.