Updated pollution multiplier minimum limits
Following Discord messages with @grilledham, updated the messages to be clearer.

Also updated some broken logic and reduced spam to admins if someone tries to set it below the minimum and it's already at the minimum.

Changes untestested. I'll test tonight or tomorrow if it passes review, not got Factorio on this laptop.